refactor: Remove `common` module
These can just be in the `mod` and marked as non-pub
