+++ prqlc/prqlc-parser/src/parser/stmt.rs
@@ -4,8 +4,8 @@ use chumsky::prelude::*;
 use itertools::Itertools;
 use semver::VersionReq;
 
-use super::common::{ctrl, ident_part, into_stmt, keyword, new_line, with_doc_comment};
 use super::expr::{expr, expr_call, ident, pipeline};
+use super::{ctrl, ident_part, into_stmt, keyword, new_line, with_doc_comment};
 use crate::lexer::lr::{Literal, TokenKind};
 use crate::parser::perror::PError;
 use crate::parser::pr::*;
