+++ prqlc/prqlc-parser/src/span.rs
@@ -15,26 +15,6 @@ pub struct Span {
     pub source_id: u16,
 }
 
-impl Span {
-    pub fn merge_opt(a: Option<Span>, b: Option<Span>) -> Option<Span> {
-        match (a, b) {
-            (None, None) => None,
-            (None, Some(s)) => Some(s),
-            (Some(s), None) => Some(s),
-            (Some(a), Some(b)) => Some(Span::merge(a, b)),
-        }
-    }
-
-    pub fn merge(a: Span, b: Span) -> Span {
-        assert_eq!(a.source_id, b.source_id);
-        Span {
-            start: usize::min(a.start, b.start),
-            end: usize::max(a.end, b.end),
-            source_id: a.source_id,
-        }
-    }
-}
-
 impl From<Span> for Range<usize> {
     fn from(a: Span) -> Self {
         a.start..a.end
